==> code/2 get_embeddings.ipynb <==
import pandas as pd
import configparser
import openai

# Load configuration file
config = configparser.ConfigParser()
config.read('../config.ini')

openai.api_key = config.get('API', 'OPENAI_API_KEY')

filename = '../data/1_first_names_raw_data.csv'

# Read the TSV file into a pandas DataFrame
df = pd.read_csv(filename, on_bad_lines='warn')

df

names = df['firstname'].tolist()
print(names[:10])


# Run once to make sure Embedding column exists
# df['Embedding'] = None

# Function to get the embedding for each firstname
def get_embedding(row):
    if row['Embedding'] is None:
        try:
            response = openai.Embedding.create(
                input=row['firstname'],
                model="text-embedding-ada-002"
            )
            embedding = response['data'][0]['embedding']
        except Exception as e:
            # Handle the exception (e.g., log the error, use a default value, etc.)
            print(f"Error in API call for '{row['firstname']}': {str(e)}")
            embedding = None  # Set a default value if the API call fails
    else:
        embedding = row['Embedding']
    
    return embedding


for index, row in df.iterrows():
    embedding_value = get_embedding(row)
    df.at[index, 'Embedding'] = embedding_value


# Assuming your DataFrame is named 'df'
not_null_count = df['Embedding'].count()

print("Number of not null values in the 'Embedding' column:", not_null_count)

# Export DataFrame to CSV file
df.to_csv('../data/2_embeddings_raw_data.csv', index=False)

df
