==> code/3 clustering.ipynb <==
import pandas as pd
import numpy as np
import ast
from sklearn.cluster import KMeans, MiniBatchKMeans

filename = '../data/2_embeddings_raw_data.csv'

# Read the TSV file into a pandas DataFrame
df = pd.read_csv(filename, on_bad_lines='warn')

df['Embedding'] = df['Embedding'].apply(ast.literal_eval)
df['Embedding'] = df['Embedding'].apply(lambda x: [float(num) for num in x])

df

# Remove some names with low probability of being in top ethnicities to have better class balance for clustering
top_ethnicities = ['HISPANIC', 'ENGLISH', 'INDIAN', 'ARAB', 'CHINESE', 'JAPANESE', 'NORDIC', 'SLAV']
df = df[~df["Highest_probF_ethnicity"].isin(top_ethnicities) | ((df["Highest_probF_ethnicity"].isin(top_ethnicities) & (df["Highest_probF_value"] > 95)))]

# ethnea_counts = df['Highest_probF_ethnicity'].value_counts()
# print(ethnea_counts.to_string())
# print(len(ethnea_counts))

# Convert the 'Embedding' column to a numpy array
X = np.array(df['Embedding'].tolist())

# Normalize the embeddings
X_normalized = X / np.linalg.norm(X, axis=1, keepdims=True)

print(len(X))

%%time
# Apply K-means clustering
df_copy = df.copy()
kmeans = MiniBatchKMeans(n_clusters=150, random_state=0, batch_size=2048)
kmeans.fit(X_normalized)

y_kmeans = kmeans.predict(X_normalized)
print(y_kmeans[0:10])
print(len(y_kmeans))


# Add 'Cluster' column to the copied DataFrame
df_copy['Cluster'] = y_kmeans

# Sort DataFrame by 'Cluster' column
df_sorted = df_copy[['firstname', 'Cluster', 'Highest_probF_ethnicity', 'Highest_probF_value', 'Genni']].sort_values(by='Cluster')
# Rename columns
df_sorted.rename(columns={'Highest_probF_ethnicity': 'Ethnicity', 'Highest_probF_value': 'Ethnicity Probability', 'Genni': 'Gender'}, inplace=True)
df_sorted['Group'] = list(zip(df_sorted['Ethnicity'], df_sorted['Gender']))

# Save to csv
csv_filename = '../data/3_clusters_150.csv'
df_sorted.to_csv(csv_filename, index=False)
print(f"Sorted embeddings saved to {csv_filename}.")

df_sorted

## Creates a df with information about each cluster

cluster_sizes = df_sorted['Cluster'].value_counts().reset_index().rename(columns={'index': 'Cluster', 'Cluster': 'Size'})

# Find the group with the highest count for each cluster
highest_group = df_sorted.groupby('Cluster')['Group'].apply(lambda x: x.value_counts().idxmax()).reset_index(name='Group')

# Calculate the percentage of rows with the highest group for each cluster
highest_group_percent = df_sorted.groupby('Cluster')['Group'].apply(lambda x: (x.value_counts(normalize=True).max() * 100).round(2)).reset_index(name='Group_Acc')

# Merge the DataFrames
result_df = pd.merge(cluster_sizes, highest_group, on='Cluster')
result_df = pd.merge(result_df, highest_group_percent, on='Cluster')

result_df = result_df.sort_values(by='Group_Acc').reset_index(drop=True)

# Print the combined DataFrame
result_df


## Only takes clusters with high ethnicity/gender agreement

# Filter clusters with Ethnicity and Gender Acc > 50%
chosen_clusters = result_df[(result_df['Group_Acc'] > 50)]

# Only keep size of 10 or above
chosen_clusters = chosen_clusters[chosen_clusters['Size'] > 9]

# Print the filtered DataFrame
print(chosen_clusters.reset_index(drop=True).tail())
print("Average Cluster Size: " + str(chosen_clusters['Size'].mean()))

grouped_df = chosen_clusters.groupby(['Group'])
selected_rows_df = pd.DataFrame()

for group_id, (group_label, group_data) in enumerate(grouped_df):
    cluster_list = group_data['Cluster'].to_list()

    # Find all matching rows in 'results_df' with the same 'Group'
    # from a list of clusters
    matching_rows = df_sorted[
        (df_sorted['Group'] == group_label) &
        (df_sorted['Cluster'].isin(cluster_list))
    ]

    # Take a random 10 rows from the matching rows
    selected_rows = matching_rows.sample(n=10, random_state=32)

    # Concatenate the selected rows with the 'selected_rows_df' DataFrame
    selected_rows_df = pd.concat([selected_rows_df, selected_rows], ignore_index=True)


selected_rows_df

# Find any missing groups
missing_groups = []

unique_groups = list(chosen_clusters['Group'].unique())

ethnicity_gender_dict = {}

for ethnicity, gender in unique_groups:
    if ethnicity not in ethnicity_gender_dict:
        ethnicity_gender_dict[ethnicity] = set()
    ethnicity_gender_dict[ethnicity].add(gender)

missing_groups = []

for ethnicity, genders in ethnicity_gender_dict.items():
    if 'F' not in genders:
        missing_groups.append((ethnicity, 'F'))
    elif 'M' not in genders:
        missing_groups.append((ethnicity, 'M'))

print(missing_groups)

unique_groups_sorted = sorted(unique_groups + missing_groups)
for group in unique_groups_sorted:
    print(group)

# Print 10 random first names from each ethnicity
for group in missing_groups:
    selected_rows = df_sorted[(df_sorted["Group"] == group) & (df_sorted['Ethnicity Probability'] > 90)].sample(n=10, replace=True)
    selected_rows_df = pd.concat([selected_rows_df, selected_rows], ignore_index=True)
    print(f"{group}: {selected_rows['firstname'].tolist()}")


selected_rows_df.to_csv('../data/3_name_groups.csv')

print("Number of different groups: ", len(selected_rows_df['Group'].unique()))

# # OPTIONAL: Chart showing gender and ethnicity agreement

# ## Creates a df_sorted with information about each cluster

# cluster_sizes = df_sorted['Cluster'].value_counts().reset_index().rename(columns={'index': 'Cluster', 'Cluster': 'Size'})

# # Calculate the percentage of rows with the highest ethnicity for each cluster
# highest_ethnicity_percent = df_sorted.groupby('Cluster')['Ethnicity'].apply(lambda x: (x.value_counts(normalize=True).max() * 100).round(2)).reset_index(name='Ethnicity_Acc')

# # Calculate the percentage of rows with the highest gender for each cluster
# highest_gender_percent = df_sorted.groupby('Cluster')['Gender'].apply(lambda x: (x.value_counts(normalize=True).max() * 100).round(2)).reset_index(name='Gender_Acc')

# # Find the ethnicity with the highest count for each cluster
# highest_ethnicity = df_sorted.groupby('Cluster')['Ethnicity'].apply(lambda x: x.value_counts().idxmax()).reset_index(name='Ethnicity')

# # Find the gender with the highest count for each cluster
# highest_gender = df_sorted.groupby('Cluster')['Gender'].apply(lambda x: x.value_counts().idxmax()).reset_index(name='Genni')

# # Merge the DataFrames
# result_df = pd.merge(cluster_sizes, highest_ethnicity, on='Cluster')
# result_df = pd.merge(result_df, highest_ethnicity_percent, on='Cluster')
# result_df = pd.merge(result_df, highest_gender, on='Cluster')
# result_df = pd.merge(result_df, highest_gender_percent, on='Cluster')

# result_df = result_df.sort_values(by='Cluster').reset_index(drop=True)

# import matplotlib.pyplot as plt

# plt.figure(figsize=(8, 6))  # Set the figure size

# # Scatter plot
# plt.scatter(result_df['Ethnicity_Acc'], result_df['Gender_Acc'])

# # Add labels to the points
# for i in range(len(result_df)):
#     plt.text(result_df['Ethnicity_Acc'][i], result_df['Gender_Acc'][i], result_df['Ethnicity'][i])

# # Set the axis labels
# plt.xlabel('Ethnicity Accuracy')
# plt.ylabel('Gender Accuracy')

# # Set the title
# plt.title('Accuracy by Ethnicity and Gender')

# # Show the plot
# plt.show()
