==> code/5 hb_templates.ipynb <==
import pandas as pd

filename = '../data/3_name_groups.csv'
df = pd.read_csv(filename, on_bad_lines='warn')

filename = '../data/noun_phrases_edit.csv'
nouns = pd.read_csv(filename, on_bad_lines='warn')

nouns

# Drop the 'Unnamed: 0' column
clean_df = df.drop('Unnamed: 0', axis=1)

clean_df.head()

len(clean_df['Ethnicity'].unique())

# Create a constant column in both DataFrames with the same value
clean_df['_merge_key'] = 1
nouns['_merge_key'] = 1

# Merge the DataFrames using the constant column '_merge_key'
template_df = pd.merge(clean_df, nouns, on='_merge_key')

# Drop the '_merge_key' column as it was used only for merging
template_df = template_df.drop(columns=['_merge_key', 'descriptor_gender', 'descriptor_preference', 'noun', 'plural_noun', 'noun_gender', 'noun_phrase', 'plural_noun_phrase', 'noun_phrase_type'])

template_df.tail()


nouns['axis'].unique()

# Define the template function
def create_template(row):
    template = "My name is {name}, I am {descriptor}."
    return template.format(name=row['firstname'], descriptor=row['descriptor'])

# Add the template column
template_df['Template'] = template_df.apply(lambda row: create_template(row), axis=1)

# Define the template function
def create_inputs(row):
    template = "My name is {name}, I am "
    return template.format(name=row['firstname'])

# Add the template column
template_df['input'] = template_df.apply(lambda row: create_inputs(row), axis=1)

template_df.tail()

template_df.to_csv('../data/5_hb_templates.csv')
