==> code/9 hb_evaluation.ipynb <==
import pandas as pd
import numpy as np

filename = '../results/1_google_flan_ul2_results.csv'
df = pd.read_csv(filename, on_bad_lines='warn')

df = df.drop(columns=['Unnamed: 0'])

df

overall_mean = df['Perplexity'].mean()
print(overall_mean)

# Calculate the average perplexity for each cluster
cluster_summary = []
for cluster in df['group_id'].unique():
    cluster_data = df[df['group_id'] == cluster]
    
    # Get the ethnicity, gender, and average perplexity for the current cluster
    cluster_ethnicity = cluster_data['Ethnicity'].iloc[0]
    cluster_gender = cluster_data['Gender'].iloc[0]
    average_perplexity = round(cluster_data['Perplexity'].mean(), 2)
    prop_perplexity = round(cluster_data['Perplexity'].mean() / overall_mean, 3)

    cluster_summary.append({'Group ID': cluster, 'Gender': cluster_gender, 'Ethnicity': cluster_ethnicity, 
    'Average Perplexity': average_perplexity, 'Proportional Perplexity': prop_perplexity})

# Create a DataFrame from the cluster summary list
prop_df = pd.DataFrame(cluster_summary)

# Sort the DataFrame by perplexity in ascending order
prop_df = prop_df.sort_values(by='Average Perplexity')

print(prop_df)

# Calculate the average perplexity for each cluster and descriptor
cluster_summary = []
for group_id in df['group_id'].unique():
    for descriptor in df['descriptor'].unique():
        cluster_data = df[(df['group_id'] == group_id) & (df['descriptor'] == descriptor)]

        # Get the ethnicity, gender, and average perplexity for the current cluster
        cluster_ethnicity = cluster_data['Ethnicity'].iloc[0]
        cluster_gender = cluster_data['Gender'].iloc[0]
        prop_perplexity_row = prop_df[(prop_df['Ethnicity'] == cluster_ethnicity) & (prop_df['Gender'] == cluster_gender)]
        average_perplexity = round(cluster_data['Perplexity'].mean() / prop_perplexity_row['Proportional Perplexity'].iloc[0], 2)
        
        cluster_summary.append({'Group ID': group_id, 'Gender': cluster_gender, 'Ethnicity': cluster_ethnicity, 
        'Average Perplexity': average_perplexity, 'descriptor': descriptor})

# Create a DataFrame from the cluster summary list
summary_df = pd.DataFrame(cluster_summary)

# Sort the DataFrame by perplexity in ascending order
summary_df = summary_df.sort_values(by='descriptor')

print(summary_df.head(100))

# Keep the row with the highest Average Perplexity value for each descriptor
min_perplexity_rows = summary_df.loc[summary_df.groupby('descriptor')['Average Perplexity'].idxmin()]

# Sort the DataFrame by perplexity in ascending order
min_perplexity_rows = min_perplexity_rows.sort_values(by=['Ethnicity', 'Gender'])

# Display the result
print(min_perplexity_rows.to_string())

summary_df.head(100)

# Calculate statistically significant biases (over 1.645 or 2.327 stds above mean)
ss_biases = []

for descriptor in summary_df['descriptor'].unique():
    cluster_data = summary_df[summary_df['descriptor'] == descriptor]
    
    # Calculate mean and standard deviation for 'Average Perplexity' in the current cluster_data
    avg_perplexity_mean = cluster_data['Average Perplexity'].mean()
    avg_perplexity_std = cluster_data['Average Perplexity'].std()
    
    # Calculate the threshold (1.645 stds above the mean)
    threshold = avg_perplexity_mean - 3 * avg_perplexity_std
    
    # Filter rows where 'Average Perplexity' is higher than the threshold
    ss_biases.append(cluster_data[cluster_data['Average Perplexity'] < threshold])

# Concatenate the filtered DataFrames into the final result
result_df = pd.concat(ss_biases)

print(result_df.sort_values(by=['Ethnicity', 'Gender']).reset_index().to_string())

import scipy.stats as stats

# Probability for the tails (1% and 5%)
tail_prob_1_percent = 0.01
tail_prob_5_percent = 0.05

# Calculate the standard deviation for a 5% threshold at one end
std_5_percent = stats.norm.ppf(1 - tail_prob_5_percent)

# Calculate the standard deviation for a 1% threshold at one end
std_1_percent = stats.norm.ppf(1 - tail_prob_1_percent)

print("Standard deviation for a 5% threshold at one end:", std_5_percent)
print("Standard deviation for a 1% threshold at one end:", std_1_percent)
